Enhance SemiDirectedPhyNetwork and MixedPhyNetwork examples with additional nodes
- Updated various functions and methods in classifications.py, derivations.py, features.py, io.py, isomorphism.py, sd_phynetwork.py, transformations.py, and layout files to include examples with additional nodes (e.g., node 100).
- Adjusted expected outputs in examples to reflect the inclusion of new nodes and edges.
- Ensured consistency across documentation and examples for better clarity and usability.

+++ src/phylozoo/core/distance/base.py
@@ -55,8 +55,7 @@ class DistanceMatrix(IOMixin):
     >>> len(dm)
     3
     >>> dm.get_distance('A', 'B')
-    1.0
-
+    np.float64(1.0)
     >>> # Default labels (0, 1, 2, ...)
     >>> dm2 = DistanceMatrix(matrix)
     >>> dm2.labels
